backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic[email]==2.5.0
qdrant-client==1.7.0
tqdm==4.66.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.1

